Move getUnitsData to a separate file
