+++ icu4c/source/i18n/i18n_uwp.vcxproj
@@ -488,6 +488,7 @@
     <ClCompile Include="unitconverter.cpp" />
     <ClCompile Include="complexunitsconverter.cpp" />
     <ClCompile Include="unitsrouter.cpp" />
+    <ClCompile Include="getunitsdata.cpp" />
     <ClCompile Include="unum.cpp" />
     <ClCompile Include="unumsys.cpp" />
     <ClCompile Include="upluralrules.cpp" />
@@ -725,6 +726,7 @@
     <ClInclude Include="unitconverter.h" />
     <ClInclude Include="complexunitsconverter.h" />
     <ClInclude Include="unitsrouter.h" />
+    <ClInclude Include="getunitsdata.h" />
   </ItemGroup>
   <ItemGroup>
     <ResourceCompile Include="i18n.rc" />
