+++ icu4c/source/i18n/unitsrouter.h
@@ -17,24 +17,6 @@
 
 U_NAMESPACE_BEGIN
 
-using icu::CharString;
-using icu::MaybeStackVector;
-using icu::MeasureUnit;
-
-struct ConversionRateInfo {
-    CharString source;
-    CharString target;
-    CharString factor;
-    CharString offset;
-};
-
-struct UnitPreference {
-    UnitPreference() : geq(1) {}
-    CharString unit;
-    double geq;
-    CharString skeleton;
-};
-
 struct ConverterPreference {
     ComplexUnitsConverter converter;
     double limit;
@@ -52,12 +34,6 @@ class U_I18N_API UnitsRouter {
   private:
     MaybeStackVector<ConverterPreference> converterPreferences_;
 };
-
-// TODO(hugo): Add a comment.
-void U_I18N_API getUnitsData(const char *outputRegion, const char *usage, const MeasureUnit &inputUnit,
-                CharString &category, MeasureUnit &baseUnit,
-                MaybeStackVector<ConversionRateInfo> &conversionInfo,
-                MaybeStackVector<UnitPreference> &unitPreferences, UErrorCode &status);
 
 U_NAMESPACE_END
